+++ test/e2e/users.test.js
@@ -1,6 +1,6 @@
 const { assert } = require('chai');
 const request = require('./request');
-const { dropCollection } = require('./db');
+const { dropCollection, createAdminToken } = require('./db');
 
 describe('User API', () => {
 
@@ -8,22 +8,8 @@ describe('User API', () => {
     before(() => dropCollection('fluffs'));
     before(() => dropCollection('users'));
     
-    const adminUser = {
-        name: 'Tina Turner',
-        password: 'Tommy',
-        roles: ['admin']
-    };
-    
-    let adminToken = null;
-    
-    before(() => {
-        return request.post('/api/auth/signup')
-            .send(adminUser)
-            .then(({ body }) => {
-                adminToken = body.token;
-                adminUser.id = body.userId;
-            });
-    });
+    let adminToken = '';
+    before(() => createAdminToken().then(t => adminToken = t));
 
     let token = '';
